+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 Next
 ----
 
+- Add support for `sphinxcontrib-mermaid`_ diagrams.
+
 2026.02.09
 ----------
 
@@ -239,3 +241,5 @@ Next
 
 2025.08.19
 ----------
+
+.. _sphinxcontrib-mermaid: https://github.com/mgaitan/sphinxcontrib-mermaid